Extract TaskSelection component
