+++ client/src/pages/assignment/AssignmentPage.tsx
@@ -5,23 +5,17 @@ import {
   CaretRightOutlined
 } from '@ant-design/icons'
 import {
-  Alert,
   Button,
-  Checkbox,
-  Col,
   DatePicker,
   Descriptions,
   Dropdown,
-  Empty,
   Form,
   Menu,
   Modal,
-  Row,
   Steps,
   TimePicker
 } from 'antd'
 import { DropdownButtonProps } from 'antd/es/dropdown/dropdown-button'
-import { CheckboxValueType } from 'antd/lib/checkbox/Group'
 import moment, { Moment, unitOfTime } from 'moment'
 import { useCallback, useState } from 'react'
 import { Route, Switch, useHistory, useRouteMatch } from 'react-router-dom'
@@ -49,7 +43,6 @@ import {
   UpdateAssignmentMutationVariables,
   useAddTasksToAssignmentMutation,
   useGetAssignmentQuery,
-  useGetTaskPoolForAddingQuery,
   useUpdateAssignmentMutation
 } from '../../services/codefreak-api'
 import { getEntityPath } from '../../services/entity-path'
@@ -65,17 +58,10 @@ import NotFoundPage from '../NotFoundPage'
 import SubmissionListPage from '../submission/SubmissionListPage'
 import TaskListPage from '../task/TaskListPage'
 import './AssignmentPage.less'
-import SortSelect from '../../components/SortSelect'
-import SearchBar from '../../components/SearchBar'
-import {
-  filterTasks,
-  TaskSortMethods,
-  TaskSortMethodNames
-} from '../../services/task'
 import { useCreateRoutes } from '../../hooks/useCreateRoutes'
 import { ShareAssignmentButton } from '../../components/ShareAssignmentButton'
 import TimeLimitTag from '../../components/time-limit/TimeLimitTag'
-import { Link } from 'react-router-dom'
+import TaskSelection from '../../components/TaskSelection'
 
 const { Step } = Steps
 
@@ -401,48 +387,8 @@ const AddTasksButton: React.FC<{
     }
   }
 
-  const [sortMethod, setSortMethod] = useState(TaskSortMethodNames[0])
-  const [filterCriteria, setFilterCriteria] = useState('')
-
-  const handleSortChange = (value: string) => setSortMethod(value)
-  const handleFilterChange = (value: string) => setFilterCriteria(value)
-
-  const sorter = (
-    <SortSelect
-      defaultValue={TaskSortMethodNames[0]}
-      values={TaskSortMethodNames}
-      onSortChange={handleSortChange}
-    />
-  )
-
-  const searchBar = (
-    <SearchBar
-      searchType="Task"
-      placeholder="by name..."
-      onChange={handleFilterChange}
-    />
-  )
-
   const taskSelection = (
-    <>
-      <Row gutter={16}>
-        <Col>{searchBar}</Col>
-        <Col>{sorter}</Col>
-      </Row>
-      <Alert
-        message={
-          'When a task from the pool is added to an assignment, an independent copy is created. ' +
-          'Editing the task in the pool will have no effect on the assignment and vice versa.'
-        }
-        style={{ marginBottom: 16, marginTop: 16 }}
-      />
-      <TaskSelection
-        value={taskIds}
-        setValue={setTaskIds}
-        sortMethod={sortMethod}
-        filterCriteria={filterCriteria}
-      />
-    </>
+    <TaskSelection selectedTaskIds={taskIds} setSelectedTaskIds={setTaskIds} />
   )
 
   return (
@@ -457,60 +403,15 @@ const AddTasksButton: React.FC<{
         title={`Add tasks to ${assignment.title}`}
         okButtonProps={{
           disabled: taskIds.length === 0,
-          loading: addTasksResult.loading
+          loading: addTasksResult.loading,
+          title:
+            taskIds.length === 0 ? "You haven't selected any tasks!" : undefined
         }}
         onOk={submit}
       >
         {taskSelection}
       </Modal>
     </>
-  )
-}
-
-const TaskSelection: React.FC<{
-  value: string[]
-  setValue: (value: string[]) => void
-  sortMethod: string
-  filterCriteria: string
-}> = props => {
-  const result = useGetTaskPoolForAddingQuery()
-
-  if (result.data === undefined) {
-    return <AsyncPlaceholder result={result} />
-  }
-
-  let taskPool = result.data.taskPool.slice()
-
-  if (taskPool.length === 0) {
-    return (
-      <Empty description={<span>Your task pool is empty.</span>}>
-        <Link to="/tasks/pool">Go to the task pool</Link> and create or import
-        your first task!
-      </Empty>
-    )
-  }
-
-  if (props.filterCriteria) {
-    taskPool = filterTasks(taskPool, props.filterCriteria)
-  }
-
-  taskPool = taskPool.sort(TaskSortMethods[props.sortMethod])
-
-  const options = taskPool.map(task => ({
-    label: task.title,
-    value: task.id
-  }))
-
-  const onChange = (value: CheckboxValueType[]) =>
-    props.setValue(value as string[])
-
-  return (
-    <Checkbox.Group
-      className="vertical-checkbox-group"
-      options={options}
-      onChange={onChange}
-      value={props.value}
-    />
   )
 }
